今日から授業

+++ src/c/speaker/memory_setting.h
@@ -1,6 +1,6 @@
 typedef struct {
   bool exit;
-  float wave[48000];
+  double wave[48000];
 } share_memory;
 
 typedef struct {
